+++ Model/BlockManager.php
@@ -7,9 +7,18 @@ use Magento\Cms\Api\Data\BlockInterface;
 use Magento\Cms\Model\BlockFactory;
 use Magento\Cms\Model\ResourceModel\Block;
 use Magento\Cms\Model\Template\FilterProvider;
+use Magento\Framework\Api\DataObjectHelper;
+use Magento\Framework\Api\SearchCriteria\CollectionProcessorInterface;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Snowdog\CmsApi\Api\BlockManagerInterface;
+use Snowdog\CmsApi\Api\Data\BlockInterfaceFactory;
+use Snowdog\CmsApi\Api\Data\BlockSearchResultsInterfaceFactory;
+use Magento\Framework\App\State;
+use Magento\Framework\App\Area;
 
+/**
+ * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
+ */
 class BlockManager implements BlockManagerInterface
 {
     /**
@@ -32,16 +41,70 @@ class BlockManager implements BlockManagerInterface
      */
     private $blockResource;
 
+    /**
+     * @var BlockSearchResultsInterfaceFactory
+     */
+    private $searchResultsFactory;
+
+    /**
+     * @var Block\CollectionFactory
+     */
+    private $blockCollectionFactory;
+
+    /**
+     * @var CollectionProcessorInterface
+     */
+    private $collectionProcessor;
+
+    /**
+     * @var BlockInterfaceFactory
+     */
+    private $blockDtoFactory;
+
+    /**
+     * @var DataObjectHelper
+     */
+    private $dataObjectHelper;
+
+    /**
+     * @var State
+     */
+    protected $appState;
+
+    /**
+     * @param BlockRepositoryInterface $blockRepository
+     * @param FilterProvider $filterProvider
+     * @param BlockFactory $blockFactory
+     * @param Block $blockResource
+     * @param Block\CollectionFactory $blockCollectionFactory
+     * @param BlockSearchResultsInterfaceFactory $searchResultsFactory
+     * @param CollectionProcessorInterface $collectionProcessor
+     * @param BlockInterfaceFactory $blockDtoFactory
+     * @param DataObjectHelper $dataObjectHelper
+     * @param State $appState
+     */
     public function __construct(
         BlockRepositoryInterface $blockRepository,
         FilterProvider $filterProvider,
         BlockFactory $blockFactory,
-        Block $blockResource
+        Block $blockResource,
+        Block\CollectionFactory $blockCollectionFactory,
+        BlockSearchResultsInterfaceFactory $searchResultsFactory,
+        CollectionProcessorInterface $collectionProcessor,
+        BlockInterfaceFactory $blockDtoFactory,
+        DataObjectHelper $dataObjectHelper,
+        State $appState
     ) {
         $this->blockRepository = $blockRepository;
         $this->filterProvider = $filterProvider;
         $this->blockFactory = $blockFactory;
         $this->blockResource = $blockResource;
+        $this->searchResultsFactory = $searchResultsFactory;
+        $this->blockCollectionFactory = $blockCollectionFactory;
+        $this->collectionProcessor = $collectionProcessor;
+        $this->blockDtoFactory = $blockDtoFactory;
+        $this->dataObjectHelper = $dataObjectHelper;
+        $this->appState = $appState;
     }
 
     /**
@@ -82,18 +145,31 @@ class BlockManager implements BlockManagerInterface
      */
     public function getList(\Magento\Framework\Api\SearchCriteriaInterface $searchCriteria)
     {
-        $blockCollection = $this->blockRepository->getList($searchCriteria);
-        $blocks = $blockCollection->getItems();
-        $parsedBlocks = [];
+        /** @var \Magento\Cms\Model\ResourceModel\Block\Collection $collection */
+        $collection = $this->blockCollectionFactory->create();
 
-        foreach ($blocks as $block) {
+        $this->collectionProcessor->process($searchCriteria, $collection);
+
+        $items = [];
+        /** @var \Magento\Cms\Model\Block $block */
+        foreach ($collection->getItems() as $block) {
             $content = $this->getBlockContentFiltered($block->getContent());
-            $parsedBlocks[] = $block->setContent($content);
+            $block->setContent($content);
+            $blockDto = $this->blockDtoFactory->create();
+            $this->dataObjectHelper->populateWithArray(
+                $blockDto,
+                $block->getData(),
+                \Snowdog\CmsApi\Api\Data\BlockInterface::class
+            );
+            $items[] = $blockDto;
         }
 
-        $blockCollection->setItems($parsedBlocks);
+        $searchResults = $this->searchResultsFactory->create();
+        $searchResults->setSearchCriteria($searchCriteria);
+        $searchResults->setItems($items);
+        $searchResults->setTotalCount(count($items));
 
-        return $blockCollection;
+        return $searchResults;
     }
 
     /**
@@ -102,7 +178,12 @@ class BlockManager implements BlockManagerInterface
      */
     private function getBlockContentFiltered($content)
     {
-        return $this->filterProvider->getBlockFilter()
-            ->filter($content);
+        $emulatedResult = $this->appState->emulateAreaCode(
+            Area::AREA_FRONTEND,
+            [$this->filterProvider->getBlockFilter(), 'filter'],
+            [$content]
+        );
+
+        return $emulatedResult;
     }
 }