+++ Model/PageManager.php
@@ -7,9 +7,18 @@ use Magento\Cms\Api\PageRepositoryInterface;
 use Magento\Cms\Model\PageFactory;
 use Magento\Cms\Model\ResourceModel\Page;
 use Magento\Cms\Model\Template\FilterProvider;
+use Magento\Framework\Api\DataObjectHelper;
+use Magento\Framework\Api\SearchCriteria\CollectionProcessorInterface;
 use Magento\Framework\Exception\NoSuchEntityException;
+use Snowdog\CmsApi\Api\Data\PageInterfaceFactory;
+use Snowdog\CmsApi\Api\Data\PageSearchResultsInterfaceFactory;
 use Snowdog\CmsApi\Api\PageManagerInterface;
+use Magento\Framework\App\State;
+use Magento\Framework\App\Area;
 
+/**
+ * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
+ */
 class PageManager implements PageManagerInterface
 {
     /**
@@ -32,16 +41,70 @@ class PageManager implements PageManagerInterface
      */
     private $pageResource;
 
+    /**
+     * @var Page\CollectionFactory
+     */
+    private $pageCollectionFactory;
+
+    /**
+     * @var PageSearchResultsInterfaceFactory
+     */
+    private $searchResultsFactory;
+
+    /**
+     * @var CollectionProcessorInterface
+     */
+    private $collectionProcessor;
+
+    /**
+     * @var PageInterfaceFactory
+     */
+    private $pageDtoFactory;
+
+    /**
+     * @var DataObjectHelper
+     */
+    private $dataObjectHelper;
+
+    /**
+     * @var State
+     */
+    protected $appState;
+
+    /**
+     * @param PageRepositoryInterface $pageRepository
+     * @param FilterProvider $filterProvider
+     * @param PageFactory $pageFactory
+     * @param Page $pageResource
+     * @param Page\CollectionFactory $pageCollectionFactory
+     * @param PageSearchResultsInterfaceFactory $searchResultsFactory
+     * @param CollectionProcessorInterface $collectionProcessor
+     * @param PageInterfaceFactory $pageDtoFactory
+     * @param DataObjectHelper $dataObjectHelper
+     * @param State $appState
+     */
     public function __construct(
         PageRepositoryInterface $pageRepository,
         FilterProvider $filterProvider,
         PageFactory $pageFactory,
-        Page $pageResource
+        Page $pageResource,
+        Page\CollectionFactory $pageCollectionFactory,
+        PageSearchResultsInterfaceFactory $searchResultsFactory,
+        CollectionProcessorInterface $collectionProcessor,
+        PageInterfaceFactory $pageDtoFactory,
+        DataObjectHelper $dataObjectHelper,
+        State $appState
     ) {
         $this->pageRepository = $pageRepository;
         $this->filterProvider = $filterProvider;
         $this->pageFactory = $pageFactory;
         $this->pageResource = $pageResource;
+        $this->pageCollectionFactory = $pageCollectionFactory;
+        $this->searchResultsFactory = $searchResultsFactory;
+        $this->collectionProcessor = $collectionProcessor;
+        $this->pageDtoFactory = $pageDtoFactory;
+        $this->dataObjectHelper = $dataObjectHelper;
+        $this->appState = $appState;
     }
 
     /**
@@ -82,18 +145,31 @@ class PageManager implements PageManagerInterface
      */
     public function getList(\Magento\Framework\Api\SearchCriteriaInterface $searchCriteria)
     {
-        $pageCollection = $this->pageRepository->getList($searchCriteria);
-        $pages = $pageCollection->getItems();
-        $parsedPages = [];
+        /** @var \Magento\Cms\Model\ResourceModel\Page\Collection $collection */
+        $collection = $this->pageCollectionFactory->create();
 
-        foreach ($pages as $page) {
+        $this->collectionProcessor->process($searchCriteria, $collection);
+
+        $items = [];
+        /** @var \Magento\Cms\Model\Page $page */
+        foreach ($collection->getItems() as $page) {
             $content = $this->getPageContentFiltered($page->getContent());
-            $parsedPages[] = $page->setContent($content);
+            $page->setContent($content);
+            $pageDto = $this->pageDtoFactory->create();
+            $this->dataObjectHelper->populateWithArray(
+                $pageDto,
+                $page->getData(),
+                \Snowdog\CmsApi\Api\Data\PageInterface::class
+            );
+            $items[] = $pageDto;
         }
 
-        $pageCollection->setItems($parsedPages);
+        $searchResults = $this->searchResultsFactory->create();
+        $searchResults->setSearchCriteria($searchCriteria);
+        $searchResults->setItems($items);
+        $searchResults->setTotalCount(count($items));
 
-        return $pageCollection;
+        return $searchResults;
     }
 
     /**
@@ -102,7 +178,12 @@ class PageManager implements PageManagerInterface
      */
     private function getPageContentFiltered($content)
     {
-        return $this->filterProvider->getPageFilter()
-            ->filter($content);
+        $emulatedResult = $this->appState->emulateAreaCode(
+            Area::AREA_FRONTEND,
+            [$this->filterProvider->getPageFilter(), 'filter'],
+            [$content]
+        );
+
+        return $emulatedResult;
     }
 }