+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] 2018-10-03
+* Widgets and content variables are parsed correctly
+* CMS Block/Pages List with store_id provided
+
 ## [1.1.0] 2018-07-23
 * Added new API endpoints to get pages/blocks by their identifier